Five conflicts, all the same shape — DATA-27's additions against the sweep's
annotation modernisation, where BOTH sides mattered:

  __init__.py       the sweep sorted __all__; load_eyegenbench goes into sorted
                    position rather than appended
  compare_source.py kept DATA-27's new `short` parameter AND the modernised
                    tuple[bool, str] return
  url_state.py      isort moved .experimental_setup / .session_keys below
                    .constants; PARAM_CORPUS and PUBLIC_DATASET_CHOICE merged
                    into the relocated import
  app.py            `from functools import partial` and `from collections.abc
                    import Callable, Mapping` are BOTH needed — git offered
                    them as either/or

  cli.py            I GOT THIS ONE WRONG FIRST. Took main's inlined
                    `_render_parser().parse_args(argv)` after grepping a
                    60-line window for reuse of `parser`. It is used at
                    cli.py:1156 — DATA-27's own --eyegenbench branch calls
                    `parser.error(...)` for a missing --eyegenbench-dataset, so
                    the inline form is a NameError the first time anyone hits
                    that flag. F821 caught it; the too-narrow search caused it.
                    The binding is restored with a comment saying why.

Then the sweep applied to this branch's own ~9,700 new lines: 12 more F821
undefined names where DATA-27 used Tuple/Optional/Dict that the sweep had
removed from those modules. Modernised to `tuple[...]` / `X | None` rather than
re-importing typing, matching the style now on main.

14 noqa directives dropped, each verified genuinely dead under the new config
(6 BLE001 — now globally ignored; 2 PLC0415 and 1 D102 — not enabled anywhere;
2 E402 and 2 N802 — enabled via extend-select, so "unused" means the rule does
not apply there). The one F401 was KEPT; only a redundant PLC0415 was trimmed
off that combined directive.

Verified: ruff 0.14.13 / 0.15.14 / 0.16.3 all clean, format clean,
`mkdocs build --strict` clean, full suite 2242 passed / 3 skipped / 3 xfailed —
identical to the pre-merge count, so neither the merge nor the sweep cost a test.

+++ scanpath_studio/update_sample_data.py
@@ -20,8 +20,8 @@ origin fails CI regardless of which script emitted it.
 from __future__ import annotations
 
 import argparse
+from collections.abc import Iterable
 from pathlib import Path
-from typing import Iterable
 
 import numpy as np
 import pandas as pd
@@ -384,7 +384,7 @@ def synthesize_raw_gaze(
     pt: list[float] = []
     n = len(trial)
     for i in range(n):
-        n_dwell = max(1, int(round(durs[i] / RAW_GAZE_SAMPLE_DT_MS)))
+        n_dwell = max(1, round(durs[i] / RAW_GAZE_SAMPLE_DT_MS))
         px.extend(xs[i] + rng.normal(0.0, RAW_GAZE_JITTER_PX, n_dwell))
         py.extend(ys[i] + rng.normal(0.0, RAW_GAZE_JITTER_PX, n_dwell))
         pt.extend(starts[i] + np.linspace(0.0, durs[i], n_dwell))
